+++ release-notes.md
@@ -8,6 +8,7 @@
 First replace all tokens in the XML file below and remove ".template" from the file name:
 - /opt/kaltura/app/deployment/updates/scripts/xml/responseProfiles/2023_02_14_add_virtual_event_response_profiles.template.xml
 - /opt/kaltura/app/deployment/updates/scripts/xml/notifications/2023_02_13_add_kafka_virtual_event_events.template.xml
+- /opt/kaltura/app/deployment/updates/scripts/xml/notifications/2023_03_08_update_kafka_kuser_notifications.template.xml
 
 To enable this feature plugin add the following to your plugins.ini file:
 - VirtualEventEventNotifications